requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
ccxt==4.4.25
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dataclasses==0.6
dateparser==1.2.0
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.6.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-yasg==1.21.8
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.67.0
gunicorn==23.0.0
h5py==3.12.1
httplib2==0.22.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
imantics==0.1.12
imgaug==0.4.0
inflection==0.5.1
iopath==0.1.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
labelme2coco==0.2.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
namex==0.0.8
narwhals==1.14.2
networkx==3.4.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pandas_ta==0.3.14b0
pathspec==0.12.1
phonenumbers==8.13.49
pillow==11.0.0
pixellib==0.7.1
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybboxes==0.1.6
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
PyMatting==1.1.12
pyparsing==3.1.4
pypng==0.20220715.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
python-binance==1.0.22
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
removebg==0.4
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
sahi==0.11.18
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sounddevice==0.5.1
sqlparse==0.5.1
streamlit==1.40.2
sympy==1.13.3
ta==0.11.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xmljson==0.2.1
yacs==0.1.8
yarl==1.17.1
